+++ src/client/index.tsx
@@ -36,7 +36,7 @@ const history = createBrowserHistory({ basename: base });
 // Clear history state (for modal routing, we want the state to be clean when refreshed)
 // OR
 // Keep the state, so we can have the same context loaded (will be slower performance)
-// history.location.state = {}
+history.location.state = {};
 
 // Grab preloaded state (that comes from the server)
 const preloadedState: RootState = (window as any).__PRELOADED_STATE__;
