+++ src/shared/App.less
@@ -13,19 +13,24 @@
     max-width: none;
   }
   &.resource-view {
-    padding: 0;
     background-color: @primary-card;
     background-image: linear-gradient(
       315deg,
       @primary-card 0%,
       @subtle-white 74%
     );
-    height: 100%;
     min-height: calc(100vh - 40px);
     transition: background-image ease-out 1s;
     .resource-details {
       .highShadow();
+      padding: 1em;
+      width: 100%;
+      max-width: 50%;
       background-color: @background-color-subtle;
+    }
+    .graph-wrapper {
+      margin-left: 1em;
+      width: 100%;
     }
   }
 }
